requirements.txt
click==7.1.2
colorama==0.4.4
fastapi==0.63.0
greenlet==1.0.0
h11==0.12.0
importlib-metadata==3.10.0
psycopg2==2.8.6
pydantic==1.8.1
python-dotenv==0.17.0
PyYAML==5.4.1
SQLAlchemy==1.4.7
starlette==0.13.6
typing-extensions==3.7.4.3
uvicorn==0.13.4
watchgod==0.7
websockets==8.1
zipp==3.4.1
